- removed unused bottom sheet functionality, placed TextWithBadge in content directory
