+++ rqes-ui-sdk/src/main/java/eu/europa/ec/rqesui/presentation/ui/component/content/ContentBottomBar.kt
@@ -1,0 +1,79 @@
+/*
+ * Copyright (c) 2023 European Commission
+ *
+ * Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the European
+ * Commission - subsequent versions of the EUPL (the "Licence"); You may not use this work
+ * except in compliance with the Licence.
+ *
+ * You may obtain a copy of the Licence at:
+ * https://joinup.ec.europa.eu/software/page/eupl
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the Licence is distributed on an "AS IS" basis, WITHOUT WARRANTIES OR CONDITIONS OF
+ * ANY KIND, either express or implied. See the Licence for the specific language
+ * governing permissions and limitations under the Licence.
+ */
+
+package eu.europa.ec.rqesui.presentation.ui.component.content
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.navigationBarsPadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import eu.europa.ec.rqesui.infrastructure.theme.values.devider
+import eu.europa.ec.rqesui.presentation.ui.component.preview.ThemeModePreviews
+import eu.europa.ec.rqesui.presentation.ui.component.utils.SPACING_LARGE
+import eu.europa.ec.rqesui.presentation.ui.component.wrap.WrapSecondaryButton
+
+@Composable
+internal fun SecondaryButtonContainerBottomBar(
+    buttonText: String,
+    onButtonClick: () -> Unit
+) {
+    Column(
+        modifier = Modifier
+            .navigationBarsPadding()
+            .fillMaxWidth()
+    ) {
+        HorizontalDivider(
+            thickness = 1.dp,
+            color = MaterialTheme.colorScheme.devider
+        )
+
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(
+                    all = SPACING_LARGE.dp
+                ),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            WrapSecondaryButton(
+                modifier = Modifier.fillMaxWidth(),
+                onClick = { onButtonClick.invoke() }
+            ) {
+                Text(
+                    text = buttonText,
+                    style = MaterialTheme.typography.labelLarge
+                )
+            }
+        }
+    }
+}
+
+@ThemeModePreviews
+@Composable
+private fun SecondaryButtonContainerBottomBarPreview() {
+    SecondaryButtonContainerBottomBar(
+        buttonText = "Close",
+        onButtonClick = {}
+    )
+}
